+++ Authentication/Configuration/ColorPalette.swift
@@ -8,37 +8,44 @@
 
 import Foundation
 
-public struct ColorPalette {
-
-    public let mainButtonDisabled: UIColor
-    public let mainButtonEnabled: UIColor
-    public let mainButtonExecuted: UIColor
+public protocol ColorPaletteType {
     
-    public let textfieldsError: UIColor
-    public let textfieldsNormal: UIColor
-    public let textfieldsSelected: UIColor
+    var mainButtonDisabled: UIColor { get }
+    var mainButtonEnabled: UIColor { get }
+    var mainButtonExecuted: UIColor { get }
     
-    public let background: UIColor
+    var textfieldsError: UIColor { get }
+    var textfieldsNormal: UIColor { get }
+    var textfieldsSelected: UIColor { get }
     
-    public let links: UIColor
-    public let labels: UIColor
-    public let textfieldText: UIColor
-    public let passwordVisibilityButtonText: UIColor
-    public let mainButtonText: UIColor
+    var background: UIColor { get }
     
-    init() {
-        mainButtonDisabled = UIColor(hexString: "#d8d8d8ff")!
-        mainButtonEnabled = UIColor(hexString: "#f5a623ff")!
-        mainButtonExecuted = UIColor(hexString: "#e78f00ff")!
-        textfieldsError = UIColor(hexString: "#d0021bff")!
-        textfieldsNormal = UIColor.clearColor()
-        textfieldsSelected = UIColor.clearColor()
-        background = UIColor(hexString: "#efefefff")!
-        links = UIColor(hexString: "#0076ffff")!
-        labels = UIColor.blackColor()
-        textfieldText = UIColor.blackColor()
-        passwordVisibilityButtonText = links
-        mainButtonText = UIColor.whiteColor()
-    }
+    var links: UIColor { get }
+    var labels: UIColor { get }
+    var textfieldText: UIColor { get }
+    var passwordVisibilityButtonText: UIColor { get }
+    var mainButtonText: UIColor { get }
     
 }
+
+public extension ColorPaletteType {
+    
+    public var mainButtonDisabled: UIColor { return UIColor(hexString: "#d8d8d8ff")! }
+    public var mainButtonEnabled: UIColor { return UIColor(hexString: "#f5a623ff")! }
+    public var mainButtonExecuted: UIColor { return UIColor(hexString: "#e78f00ff")! }
+    
+    public var textfieldsError: UIColor { return UIColor(hexString: "#d0021bff")! }
+    public var textfieldsNormal: UIColor { return UIColor.clearColor() }
+    public var textfieldsSelected: UIColor { return UIColor.clearColor() }
+    
+    public var background: UIColor { return UIColor(hexString: "#efefefff")! }
+    
+    public var links: UIColor { return UIColor(hexString: "#0076ffff")! }
+    public var labels: UIColor { return UIColor.blackColor() }
+    public var textfieldText: UIColor { return UIColor.blackColor() }
+    public var passwordVisibilityButtonText: UIColor { return UIColor(hexString: "#0076ffff")! }
+    public var mainButtonText: UIColor { return UIColor.whiteColor() }
+    
+}
+
+public struct DefaultColorPalette: ColorPaletteType { }
