+++ Authentication/SignUp/View/SignupViewConfiguration.swift
@@ -10,19 +10,19 @@ import Foundation
 
 public protocol SignupViewConfigurationType {
     
-    var colourPalette: ColorPalette { get }
-    var fontPalette: FontPalette { get }
+    var colourPalette: ColorPaletteType { get }
+    var fontPalette: FontPaletteType { get }
     
 }
 
 public final class DefaultSignupViewConfiguration: SignupViewConfigurationType {
     
-    public let colourPalette: ColorPalette
-    public let fontPalette: FontPalette
+    public let colourPalette: ColorPaletteType
+    public let fontPalette: FontPaletteType
     
     public init() {
-        colourPalette = ColorPalette()
-        fontPalette = FontPalette()
+        colourPalette = DefaultColorPalette()
+        fontPalette = DefaultFontPalette()
     }
     
 }
